+++ lib/api_form.dart
@@ -730,6 +730,7 @@ class APIFormField {
 
         return ListTile(
           title: Text(item.partName),
+          subtitle: Text(item.locationPathString),
           leading: InvenTreeAPI().getThumbnail(item.partThumbnail),
           trailing: Text(item.quantityString()),
         );
@@ -1021,17 +1022,16 @@ Future<void> launchApiForm(
   APIFormWidgetState? formHandler,
   IconData icon = TablerIcons.device_floppy,
 }) async {
-  showLoadingOverlay();
-
   // List of fields defined by the server
   Map<String, dynamic> serverFields = {};
 
   if (url.isNotEmpty) {
+    showLoadingOverlay();
     var options = await InvenTreeAPI().options(url);
+    hideLoadingOverlay();
 
     // Invalid response from server
     if (!options.isValid()) {
-      hideLoadingOverlay();
       return;
     }
 
@@ -1040,8 +1040,6 @@ Future<void> launchApiForm(
     if (serverFields.isEmpty) {
       // User does not have permission to perform this action
       showSnackIcon(L10().response403, icon: TablerIcons.user_x);
-
-      hideLoadingOverlay();
       return;
     }
   }
@@ -1084,6 +1082,8 @@ Future<void> launchApiForm(
     }
     formFields.add(field);
   }
+
+  showLoadingOverlay();
 
   // Grab existing data for each form field
   for (var field in formFields) {
@@ -1440,7 +1440,7 @@ class APIFormWidgetState extends State<APIFormWidget> {
     // Perhaps we just want to process the data?
     if (widget.url.isEmpty) {
       // Hide the form
-      handleSuccess(data, {});
+      handleSuccess(data, data);
       return;
     }