+++ lib/preferences.dart
@@ -33,9 +33,9 @@ const String INV_LABEL_DEFAULT_PRINTER = "defaultLabelPrinter";
 const String INV_LABEL_DEFAULT_PLUGIN = "defaultLabelPlugin";
 
 // Part settings
-const String INV_PART_SHOW_PARAMETERS = "partShowParameters";
 const String INV_PART_SHOW_BOM = "partShowBom";
 const String INV_PART_SHOW_PRICING = "partShowPricing";
+const String INV_PART_SHOW_REQUIREMENTS = "partShowRequirements";
 
 // Stock settings
 const String INV_STOCK_SHOW_HISTORY = "stockShowHistory";