+++ lib/settings/part_settings.dart
@@ -13,12 +13,9 @@ class InvenTreePartSettingsWidget extends StatefulWidget {
 class _InvenTreePartSettingsState extends State<InvenTreePartSettingsWidget> {
   _InvenTreePartSettingsState();
 
-  bool partShowParameters = true;
   bool partShowBom = true;
   bool partShowPricing = true;
-  bool stockShowHistory = false;
-  bool stockShowTests = false;
-  bool stockConfirmScan = false;
+  bool partShowRequirements = false;
 
   @override
   void initState() {
@@ -28,10 +25,6 @@ class _InvenTreePartSettingsState extends State<InvenTreePartSettingsWidget> {
   }
 
   Future<void> loadSettings() async {
-    partShowParameters = await InvenTreeSettingsManager().getBool(
-      INV_PART_SHOW_PARAMETERS,
-      true,
-    );
     partShowBom = await InvenTreeSettingsManager().getBool(
       INV_PART_SHOW_BOM,
       true,
@@ -40,16 +33,8 @@ class _InvenTreePartSettingsState extends State<InvenTreePartSettingsWidget> {
       INV_PART_SHOW_PRICING,
       true,
     );
-    stockShowHistory = await InvenTreeSettingsManager().getBool(
-      INV_STOCK_SHOW_HISTORY,
-      false,
-    );
-    stockShowTests = await InvenTreeSettingsManager().getBool(
-      INV_STOCK_SHOW_TESTS,
-      true,
-    );
-    stockConfirmScan = await InvenTreeSettingsManager().getBool(
-      INV_STOCK_CONFIRM_SCAN,
+    partShowRequirements = await InvenTreeSettingsManager().getBool(
+      INV_PART_SHOW_REQUIREMENTS,
       false,
     );
 
@@ -68,23 +53,6 @@ class _InvenTreePartSettingsState extends State<InvenTreePartSettingsWidget> {
       body: Container(
         child: ListView(
           children: [
-            ListTile(
-              title: Text(L10().parameters),
-              subtitle: Text(L10().parametersSettingDetail),
-              leading: Icon(TablerIcons.list),
-              trailing: Switch(
-                value: partShowParameters,
-                onChanged: (bool value) {
-                  InvenTreeSettingsManager().setValue(
-                    INV_PART_SHOW_PARAMETERS,
-                    value,
-                  );
-                  setState(() {
-                    partShowParameters = value;
-                  });
-                },
-              ),
-            ),
             ListTile(
               title: Text(L10().bom),
               subtitle: Text(L10().bomEnable),
@@ -116,54 +84,19 @@ class _InvenTreePartSettingsState extends State<InvenTreePartSettingsWidget> {
                 },
               ),
             ),
-            Divider(),
             ListTile(
-              title: Text(L10().stockItemHistory),
-              subtitle: Text(L10().stockItemHistoryDetail),
-              leading: Icon(TablerIcons.history),
+              title: Text(L10().partRequirements),
+              subtitle: Text(L10().partRequirementsSettingDetail),
+              leading: Icon(TablerIcons.list),
               trailing: Switch(
-                value: stockShowHistory,
+                value: partShowRequirements,
                 onChanged: (bool value) {
                   InvenTreeSettingsManager().setValue(
-                    INV_STOCK_SHOW_HISTORY,
+                    INV_PART_SHOW_REQUIREMENTS,
                     value,
                   );
                   setState(() {
-                    stockShowHistory = value;
-                  });
-                },
-              ),
-            ),
-            ListTile(
-              title: Text(L10().testResults),
-              subtitle: Text(L10().testResultsDetail),
-              leading: Icon(TablerIcons.test_pipe),
-              trailing: Switch(
-                value: stockShowTests,
-                onChanged: (bool value) {
-                  InvenTreeSettingsManager().setValue(
-                    INV_STOCK_SHOW_TESTS,
-                    value,
-                  );
-                  setState(() {
-                    stockShowTests = value;
-                  });
-                },
-              ),
-            ),
-            ListTile(
-              title: Text(L10().confirmScan),
-              subtitle: Text(L10().confirmScanDetail),
-              leading: Icon(TablerIcons.qrcode),
-              trailing: Switch(
-                value: stockConfirmScan,
-                onChanged: (bool value) {
-                  InvenTreeSettingsManager().setValue(
-                    INV_STOCK_CONFIRM_SCAN,
-                    value,
-                  );
-                  setState(() {
-                    stockConfirmScan = value;
+                    partShowRequirements = value;
                   });
                 },
               ),