+++ lib/widget/parameter_widget.dart
@@ -1,8 +1,12 @@
 import "package:flutter/material.dart";
+import "package:flutter_tabler_icons/flutter_tabler_icons.dart";
+import "package:inventree/api.dart";
+import "package:inventree/app_colors.dart";
 import "package:inventree/inventree/model.dart";
+import "package:inventree/inventree/parameter.dart";
 
 import "package:inventree/l10.dart";
-import "package:inventree/inventree/part.dart";
+import "package:inventree/widget/link_icon.dart";
 import "package:inventree/widget/paginator.dart";
 import "package:inventree/widget/progress.dart";
 import "package:inventree/widget/refreshable_state.dart";
@@ -10,16 +14,18 @@ import "package:inventree/widget/refreshable_state.dart";
 /*
  * Widget for displaying a list of parameters associated with a given Part instance
  */
-class PartParameterWidget extends StatefulWidget {
-  const PartParameterWidget(this.part);
+class ParameterWidget extends StatefulWidget {
+  const ParameterWidget(this.modelType, this.modelId, this.editable) : super();
 
-  final InvenTreePart part;
+  final String modelType;
+  final int modelId;
+  final bool editable;
 
   @override
   _ParameterWidgetState createState() => _ParameterWidgetState();
 }
 
-class _ParameterWidgetState extends RefreshableState<PartParameterWidget> {
+class _ParameterWidgetState extends RefreshableState<ParameterWidget> {
   _ParameterWidgetState();
 
   @override
@@ -34,9 +40,16 @@ class _ParameterWidgetState extends RefreshableState<PartParameterWidget> {
 
   @override
   Widget getBody(BuildContext context) {
-    Map<String, String> filters = {"part": widget.part.pk.toString()};
+    Map<String, String> filters = {
+      "model_type": widget.modelType,
+      "model_id": widget.modelId.toString(),
+    };
 
-    return Column(children: [Expanded(child: PaginatedParameterList(filters))]);
+    return Column(
+      children: [
+        Expanded(child: PaginatedParameterList(filters, widget.editable)),
+      ],
+    );
   }
 }
 
@@ -44,8 +57,10 @@ class _ParameterWidgetState extends RefreshableState<PartParameterWidget> {
  * Widget for displaying a paginated list of Part parameters
  */
 class PaginatedParameterList extends PaginatedSearchWidget {
-  const PaginatedParameterList(Map<String, String> filters)
+  const PaginatedParameterList(Map<String, String> filters, this.editable)
     : super(filters: filters);
+
+  final bool editable;
 
   @override
   String get searchTitle => L10().parameters;
@@ -75,7 +90,7 @@ class _PaginatedParameterState
     int offset,
     Map<String, String> params,
   ) async {
-    final page = await InvenTreePartParameter().listPaginated(
+    final page = await InvenTreeParameter().listPaginated(
       limit,
       offset,
       filters: params,
@@ -84,7 +99,7 @@ class _PaginatedParameterState
     return page;
   }
 
-  Future<void> editParameter(InvenTreePartParameter parameter) async {
+  Future<void> editParameter(InvenTreeParameter parameter) async {
     // Checkbox values are handled separately
     if (parameter.is_checkbox) {
       return;
@@ -101,7 +116,7 @@ class _PaginatedParameterState
 
   @override
   Widget buildItem(BuildContext context, InvenTreeModel model) {
-    InvenTreePartParameter parameter = model as InvenTreePartParameter;
+    InvenTreeParameter parameter = model as InvenTreeParameter;
 
     String title = parameter.name;
 
@@ -116,7 +131,7 @@ class _PaginatedParameterState
           ? Switch(
               value: parameter.as_bool,
               onChanged: (bool value) {
-                if (parameter.canEdit) {
+                if (widget.editable) {
                   showLoadingOverlay();
                   parameter.update(values: {"data": value.toString()}).then((
                     value,
@@ -127,14 +142,51 @@ class _PaginatedParameterState
                 }
               },
             )
-          : Text(parameter.value),
+          : LargeText(parameter.value),
       onTap: parameter.is_checkbox
           ? null
           : () async {
-              if (parameter.canEdit) {
+              if (widget.editable) {
                 editParameter(parameter);
               }
             },
     );
   }
+}
+
+/*
+ * Return a ListTile to display parameters for the specified model
+ */
+ListTile? ShowParametersItem(
+  BuildContext context,
+  String modelType,
+  int modelId,
+  int parameterCount,
+  bool editable,
+) {
+  // Note: Currently cannot add parameters from the app,
+  // So, if there are no parameters, do not show the item
+  if (parameterCount == 0) {
+    return null;
+  }
+
+  if (!InvenTreeAPI().supportsModernParameters) {
+    return null;
+  }
+
+  return ListTile(
+    title: Text(L10().parameters),
+    leading: Icon(TablerIcons.list_details, color: COLOR_ACTION),
+    trailing: LinkIcon(
+      text: parameterCount > 0 ? parameterCount.toString() : null,
+    ),
+    onTap: () {
+      Navigator.push(
+        context,
+        MaterialPageRoute(
+          builder: (context) => ParameterWidget(modelType, modelId, editable),
+        ),
+      );
+    },
+  );
 }