+++ lib/widget/stock/stock_list.dart
@@ -75,8 +75,13 @@ class _PaginatedStockItemListState
         "help_text": L10().filterInStockDetail,
         "tristate": true,
       },
+      "active": {
+        "default": true,
+        "label": L10().filterActive,
+        "help_text": L10().filterActiveDetail,
+      },
       "cascade": {
-        "default": false,
+        "default": true,
         "label": L10().includeSublocations,
         "help_text": L10().includeSublocationsDetail,
         "tristate": false,